% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficientes.R
\docType{package}
\name{coeficientes}
\alias{coeficientes}
\title{Coeficientes do modelo estimado}
\description{
Funcao para pegar os coeficientes estimados no modelo linear de regressao - linmod.
}
\examples{
x = cbind(1, dados2$x, dados2$z,dados2$w)
y = dados2$y
coeficientes(x,y)
}
\author{
Jonathan S. Matias
}
