% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{dados}
\alias{dados}
\title{Equacao do Consumo}
\format{
Um data frame com 54 linhas and 2 variaveis sendo:
\itemize{
item x: renda disponivel;
item y: consumo.
}
}
\source{
Bureau of Economic Analysis, Departamento de Comércio dos EUA: http://www.bea.doc.gov/bea/dn1.htm
}
\description{
Despesa real de consumo e renda real para os Estados Unidos no periodo de 1947 a 2000.
}
\author{
Jonathan S. Matias \email{onathanestatistica@gmail.com }
}
\keyword{consumo,}
\keyword{disponivel}
\keyword{renda}
