% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados2.R
\docType{data}
\name{dados2}
\alias{dados2}
\title{Equacao do Consumo multivariada}
\format{
Um data frame com 54 linhas and 4 variaveis sendo:
\itemize{
item x: renda disponivel;
item y: consumo;
item z: riqueza real;
item w: taxa de juros
}
}
\source{
Bureau of Economic Analysis, Departamento de Comércio dos EUA: http://www.bea.doc.gov/bea/dn1.htm
}
\description{
Despesa real de consumo, renda real, riqueza real e taxas reais de juros para os Estados Unidos no período 1947-2000.
}
\author{
Jonathan S. Matias \email{onathanestatistica@gmail.com }
}
\keyword{consumo,}
\keyword{disponivel,}
\keyword{e}
\keyword{juros}
\keyword{renda}
\keyword{riqueza}
