% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interconf.R
\docType{package}
\name{interconf}
\alias{interconf}
\title{Funcao para calcular os intervalos de confianca dos parametros}
\description{
Esta funcao calcula o intervalo de confianca em torno nos parametros estimados considerando o nivel de confianca dado.
}
\examples{
conf <- interconf(cbind(1,dados2$x,dados2$y,dados2$z,dados2$w),dados2$y,.0975)
conf

}
