% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste.R
\docType{package}
\name{linmod.defaut}
\alias{linmod.defaut}
\title{limod_defaut}
\description{
Funcao limod para adicionar o metodo que estara disponivel no default.
}
\examples{
reg1 <- linmod(cbind(1, dados2$x, dados2$z, dados2$w), dados$y)
reg1

}
\author{
Jonathan S. Matias
}
