% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste.R
\docType{package}
\name{linmod.formula}
\alias{linmod.formula}
\alias{jonathan_ajuste3}
\title{limod_formula}
\description{
Funcao que tem como objetivo estabelecer a formula basica que deve ser utilizada na funcao a fim de estimar o modelo.
}
\author{
Jonathan S. Matias
}
