% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumario.R
\docType{package}
\name{print.linmod}
\alias{print.linmod}
\alias{jonathan_sumario}
\title{funcao linmod_class}
\description{
Funcao que fixa a classe de retorno dos objetos "linmod".
}
\author{
Jonathan S. Matias
}
