% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumario.R
\docType{package}
\name{print.summary}
\alias{print.summary}
\title{Sumario_adicional}
\description{
Funcao com objetivo de exibier uma matriz com melhor arredondamento e visualmente melhor.
}
\author{
Jonathan S. Matias
}
