% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumario.R
\docType{package}
\name{sumario}
\alias{sumario}
\title{Sumario}
\description{
Modelo sumarizado em 4 colunas contendo as estimativas, erro padrao, estatistica t e p-valor.
}
\examples{
x = cbind(1, dados2$x, dados2$z,dados2$w)
y = dados2$y
mod = linmod(x,y,z,w)
summary(mod)
}
\author{
Jonathan S. Matias
}
