% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumario.R
\docType{package}
\name{summary.linmod}
\alias{summary.linmod}
\alias{jonathan_sumario1}
\title{sumario_princial
The parameter estimates of a statistical}
\description{
Modelo sumarizado em 4 colunas contendo as estimativas, erro padrao, estatistica t e p-valor.
}
\author{
Jonathan S. Matias
}
