% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\docType{package}
\name{vcov}
\alias{vcov}
\title{Funcao variancia_covariancias}
\description{
Funcao que tem como objetivo estimar a matriz de variancia e covariancias do modelo com base no linmod
}
\examples{
x = cbind(1, dados2$x, dados2$z,dados2$w)
y = dados$y
vcov(x,y)

}
