% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interconf.R
\docType{package}
\name{interconf}
\alias{interconf}
\alias{jonathan_interconf}
\title{Funcao para calcular os intervalos de confianca dos parametros}
\description{
Esta funcao calcula o intervalo de confianca em torno nos parametros estimados considerando o nivel de confianca dado.
}
