% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste.R
\docType{package}
\name{linmod.defaut}
\alias{linmod.defaut}
\alias{jonathan_ajuste2}
\title{limod_defaut}
\description{
Funcao limod para adicionar o metodo que estara disponivel no default.
}
\author{
Jonathan S. Matias
}
