% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste.R
\docType{package}
\name{linmod.defaut}
\alias{linmod.defaut}
\title{limod_defaut}
\description{
Funcao limod para adicionar o metodo que estara disponivel no default.
}
\examples{
reg3 <- linmod.default(cbind(1, dados$x), dados$y)
reg3

}
\author{
Jonathan S. Matias
}
