% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste.R
\docType{package}
\name{linmodEst}
\alias{linmodEst}
\alias{jonathan_ajuste}
\title{funcao de ajuste do modelo}
\description{
Funcoes que tem como objetivo ajusta o modelo com base nos Minimos Quadrados e estimar todos os parametros.
}
\examples{
reg <- linmodEst(cbind(1, dados$x), dados$y)
reg

}
\author{
Jonathan S. Matias
}
