% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumario.R
\docType{package}
\name{print.linmod}
\alias{print.linmod}
\alias{#'}
\alias{@examples}
\alias{x}
\alias{=}
\alias{cbind(1,}
\alias{dados2$x,}
\alias{dados2$z,dados2$w)}
\alias{y}
\alias{dados2$y}
\alias{mod}
\alias{linmod(x,y,z,w)}
\alias{print(mod)}
\title{Coeficientes}
\description{
Funcao que ira fazer o print dos coeficientes do modelo estimado usando a funcao linmod.
}
\author{
Jonathan S. Matias
}
